Fixed Quality Settings Bug
Fixed a bug where a blank playlist is generated when a user attempts to load a stream which does not have their preferred quality available. Addon now checks that the preferred quality is available for the stream being requested and defaults to the playlist provided by Twitch's API in the case that multiple qualities are not available (eg: unpartnered small streams).